split model parallel models and modules to enc/dec, matching non mp structure
